Add test code subtracting instrumental pol from Q and U.
Initial code measuring the median in a narrow annulus with
the disk masked out above/below the 1-sigma level, scaling that
median by the I radial median, and subtracting it from the data.
Do this for both Q and U, and plot. Produces very similar results
for radial Stokes as the original method that fits quadrupole to
U_phi, which still happens.